Generalize Grid Item. Thumbnails Linkable.
