+++ vf-react/src/static/css/style.css
@@ -41,6 +41,10 @@ ul.nav a:hover {
       transform: translateX(-50%) translateY(-50%);
 }
 
+a:focus, a:hover {
+  text-decoration: none;
+}
+
 
 
 #logohome {
